Factored out the advanced speed test interval
